pyproject.toml
[tool.poetry]
name = "snapOCR"
version = "0.0.1"
description = "A tool for snapping a capture of a region of interest and extracting the text"
authors = ["skorm11x <kosik.chris.j@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
pytesseract = "^0.3.13"
pillow = "^10.4.0"
pyperclip = "^1.9.0"
pyqt5 = "^5.15.11"
pylint = "^3.2.6"
pyinstaller = "^6.10.0"
black = "^24.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
